Extracted IgCacheFactoryBase from IgCacheFactory.
